pyproject.toml
[tool.poetry]
name = "trocadalho-generator-bot"
version = "0.1.0"
description = ""
authors = ["b0rba <borbalves.matheus@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
python-dotenv = "^0.19.2"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
